Side Bag - August 12

+++ Fulfil.Dispense/include/Fulfil.Dispense/commands/pre_side_dispense/pre_side_dispense_response.h
@@ -26,7 +26,7 @@ namespace fulfil::dispense::commands {
     public:
         explicit PreSideDispenseResponse(std::shared_ptr<std::string> request_id, 
                                          std::shared_ptr<std::string> primary_key_id,
-                                         std::shared_ptr<std::vector<std::vector<int>>> occupancy_map,
+                                         std::shared_ptr<std::vector<std::vector<float>>> occupancy_map,
                                          SideDispenseErrorCodes success_code, 
                                          std::string error_description=std::string(""));
 
@@ -34,7 +34,7 @@ namespace fulfil::dispense::commands {
         std::shared_ptr<std::string> get_command_id() override;
         std::shared_ptr<std::string> dispense_payload() override;
         std::shared_ptr<std::string> primary_key_id;
-        std::shared_ptr<std::vector<std::vector<int>>> occupancy_map;
+        std::shared_ptr<std::vector<std::vector<float>>> occupancy_map;
         SideDispenseErrorCodes success_code{SideDispenseErrorCodes::Success};
         std::string error_description{""};
     };
